Refactor into separate hook
